refactor: menu의 menuGroup 참조를 간접참조로 수정하라
